% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_low_quality.R
\name{filter_quality_deezer_tracks}
\alias{filter_quality_deezer_tracks}
\title{Filter low quality results for \emph{Deezer} Tracks}
\usage{
filter_quality_deezer_tracks(frame, track_threshold, firstartist_threshold)
}
\arguments{
\item{frame}{Data frame containing Deezer tracks variables.}

\item{track_threshold}{Threshold for Deezer track quality. Ranging between 0 and 1.}

\item{firstartist_threshold}{Threshold for Deezer artist quality. Ranging between 0 and 1.}
}
\value{
Updated input frame with all Deezer track information set to NA where at least one threshold was not met.
}
\description{
Filter low quality results for \emph{Deezer} Tracks
}
