% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_low_quality.R
\name{filter_quality_musicbrainz_acousticbrainz_all}
\alias{filter_quality_musicbrainz_acousticbrainz_all}
\title{Filter low quality results for all \emph{Musicbrainz} and \emp{Acousticbrainz} Information}
\usage{
filter_quality_musicbrainz_acousticbrainz_all(
  frame,
  track_threshold,
  album_threshold,
  artist_threshold
)
}
\arguments{
\item{frame}{Data frame containing all Musicbrainz and optional also Acousticbrainz variables.}

\item{track_threshold}{Threshold for Musicbrainz track quality. Ranging between 0 and 1.}

\item{album_threshold}{Threshold for Musicbrainz track quality. Ranging between 0 and 1.}

\item{artist_threshold}{Threshold for Musicbrainz track quality. Ranging between 0 and 1.}
}
\value{
Updated input frame with all Musicbrainz (and Acousticbrainz) information set to NA where at least one threshold was not met. For album and track, artist quality will also be considered.
}
\description{
Filter low quality results for all \emph{Musicbrainz} and \emp{Acousticbrainz} Information
}
