% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_low_quality.R
\name{filter_quality_musicbrainz_acousticbrainz_tracks}
\alias{filter_quality_musicbrainz_acousticbrainz_tracks}
\title{Filter low quality results for \emph{Musicbrainz} and \emph{Acousticbrainz} Tracks}
\usage{
filter_quality_musicbrainz_acousticbrainz_tracks(frame, track_threshold)
}
\arguments{
\item{frame}{Data frame containing Musicbrainz track variables.}

\item{track_threshold}{Threshold for Musicbrainz track quality. Ranging between 0 and 1.}
}
\value{
Updated input frame with all Musicbrainz and Acousticbrainz track information set to NA where at least one threshold was not met.
}
\description{
Filter low quality results for \emph{Musicbrainz} and \emph{Acousticbrainz} Tracks
}
