% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_low_quality.R
\name{filter_quality_musicbrainz_albums}
\alias{filter_quality_musicbrainz_albums}
\title{Filter low quality results for \emph{Musicbrainz} Albums}
\usage{
filter_quality_musicbrainz_albums(frame, album_threshold)
}
\arguments{
\item{frame}{Data frame containing Musicbrainz album variables.}

\item{album_threshold}{Threshold for Musicbrainz album quality. Ranging between 0 and 1.}
}
\value{
Updated input frame with all Musicbrainz album information set to NA where at least one threshold was not met.
}
\description{
Filter low quality results for \emph{Musicbrainz} Albums
}
