% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_low_quality.R
\name{filter_quality_musicbrainz_artists}
\alias{filter_quality_musicbrainz_artists}
\title{Filter low quality results for \emph{Musicbrainz} Artists}
\usage{
filter_quality_musicbrainz_artists(frame, artist_threshold)
}
\arguments{
\item{frame}{Data frame containing Musicbrainz artist variables.}

\item{artist_threshold}{Threshold for Musicbrainz artist quality. Ranging between 0 and 1.}
}
\value{
Updated input frame with all Musicbrainz artist information set to NA where at least one threshold was not met.
}
\description{
Filter low quality results for \emph{Musicbrainz} Artists
}
