% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_albums_discogs.R
\name{get_albums_discogs}
\alias{get_albums_discogs}
\title{Get \emph{Discogs} Album Genres and Style}
\usage{
get_albums_discogs(
  input,
  dc_pass,
  album_threshold = 0.8,
  artist_threshold = 0.8
)
}
\arguments{
\item{input}{Data Frame containing the following columns:
\itemize{
\item \code{album.s.id} \cr
with \emph{Spotify} album id,
\item \code{album.s.title} \cr
with \emph{Spotify} album title,
\item \code{track.s.firstartist.name} \cr
with \emph{Spotify} artist name.
}
It is advisable to first run \code{\link{get_all_spotify}} before running this command,
in order to have all the necessary information.}

\item{dc_pass}{Discogs Authentification. Please refer to \url{https://www.discogs.com/developers/#page:authentication,header:authentication-discogs-auth-flow} for details.}

\item{threshold}{Floating point number between 0 and 1 indicating which elements to keep that were found using a fuzzy search.
The values correspond to the string similarity (1 - Jaro-Winkler distance) between the searched artist / album and the found name / title on \emph{Spotify}. \emph{spotilink} will only keep results where the artist name as well as the album title surpass the threshold.}
}
\description{
Get \emph{Discogs} Album Genres and Style
}
