% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_albums_musicbrainz.R
\name{get_albums_musicbrainz}
\alias{get_albums_musicbrainz}
\title{Get \emph{MusicBrainz} Album Information}
\usage{
get_albums_musicbrainz(
  input,
  album_threshold = 0.8,
  firstartist_threshold = 0.8
)
}
\arguments{
\item{input}{Data Frame containing the following columns:
\itemize{
\item \code{album.s.id} \cr
with \emph{Spotify} album id,
\item \code{album.s.title} \cr
with \emph{Spotify} album title,
\item \code{album.s.upc} \cr
with album UPCs from \emph{Spotify}.
}
It is advisable to first run \code{\link{get_albums_spotify}} before running this command,
in order to have all the necessary information.}

\item{album_threshold, firstartist_threshold}{Floating point number between 0 and 1 indicating which elements to keep that were found using a fuzzy search.
The values correspond to the string similarity (1 - Jaro-Winkler distance) between the album title / artist name on \emph{Spotify} and the found album title / artist name on \emph{Musicbrainz}.}
}
\value{
Data Frame with added information from the \emph{MusicBrainz} API using the  \pkg{spotilink} naming convention.
}
\description{
Retrieve a data frame containing information from the \emph{MusicBrainz} API.
The result contains information on albums.
The \pkg{spotilink} naming convention is used.
}
\examples{
data <- data.frame(album.s.id = c('2rlWvQ1GuTOSkyNpmcoaMC'),
                  album.s.title = c('Bach: Motets'),
                  album.s.firstartist.name = c('Johann Sebastian Bach'),
                  album.s.releaseyear = c(2012),
                  album.s.upc = c('843183071623'))

get_albums_musicbrainz(data)
}
