% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_albums_spotify.R
\name{get_albums_spotify}
\alias{get_albums_spotify}
\title{Get \emph{Spotify} API Information on Albums}
\usage{
get_albums_spotify(input, pass)
}
\arguments{
\item{input}{Data Frame containing a column \code{album.s.id} with \emph{Spotify} album ids.}

\item{pass}{Character Vector containing two entries: \emph{Client ID} and \emph{Client secret}. See \url{https://developer.spotify.com/documentation/web-api/concepts/authorization} for details.}
}
\value{
Data Frame with added information from the \emph{Spotify} Web API using the \pkg{spotilink} naming convention.
}
\description{
Retrieve a data frame containing information from the \emph{Spotify} API.
The result contains information on the albums using the  \pkg{spotilink} naming convention.
}
