% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_deezer.R
\name{get_all_deezer}
\alias{get_all_deezer}
\title{Get \emph{Deezer} Track Information}
\usage{
get_all_deezer(
  input,
  track_threshold = 0.8,
  album_threshold = 0.8,
  firstartist_threshold = 0.8
)
}
\arguments{
\item{input}{Data Frame containing the following columns:
\itemize{
\item\code{track.s.id} \cr
with \emph{Spotify} track ids,
\item \code{track.s.title} \cr
with \emph{Spotify} track name,
\item \code{track.s.firstartist.name} \cr
with \code{Spotify} name of first artist,
\item \code{album.s.title} \cr
with album title from \emph{Spotify}
}
It is advisable to first run \code{\link{get_tracks_spotify}} before running this command,
in order to have all the necessary information.}

\item{track_threshold, album_threshold, firstartist_threshold}{Floating point number between 0 and 1 indicating which elements to keep that were found using a fuzzy search.
The values correspond to the string similarity (1 - Jaro-Winkler distance) between the track / album title or artist name  on \emph{Spotify} and the found track / album title or artist name on \emph{Deezer}.}
}
\value{
Data Frame with added information from the \emph{Deezer} API using the  \pkg{spotilink} naming convention.
}
\description{
Retrieve a data frame containing information from the \emph{Deezer} API.
The result contains information on tracks, albums, and artists as inferred by the track search. Threfore, only information on the percentage of linked tracks will be considered.
The \pkg{spotilink} naming convention is used.
}
