% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_musicbrainz.R
\name{get_all_musicbrainz}
\alias{get_all_musicbrainz}
\title{Get \emph{MusicBrainz} Information}
\usage{
get_all_musicbrainz(input)
}
\arguments{
\item{input}{Data Frame containing the following columns:
\itemize{
\item\code{track.s.id} \cr
with \emph{Spotify} track ids,
\item \code{track.s.title} \cr
with \emph{Spotify} track name,
\item \code{track.s.isrc} \cr
with track ISRCs from \emph{Spotify},
\item \code{album.s.id} \cr
with \emph{Spotify} album id,
\item \code{album.s.title} \cr
with \emph{Spotify} album title,
\item \code{album.s.upc} \cr
with album UPCs from \emph{Spotify}.
\item \code{artist.s.id} \cr
with \emph{Spotify} artist id,
\item \code{artist.s.name} \cr
with \emph{Spotify} artist name.
}
It is advisable to first run \code{\link{get_all_spotify}} before running this command,
in order to have all the necessary information.}
}
\value{
Data Frame with added information from the \emph{MusicBrainz} API using the  \pkg{spotilink} naming convention.
}
\description{
Retrieve a data frame containing information from the \emph{MusicBrainz} API.
The result contains information on the tracks as well as corresponding albums and artists.
The \pkg{spotilink} naming convention is used.
}
