% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_spotify.R
\name{get_all_spotify}
\alias{get_all_spotify}
\title{Get \emph{Spotify} API Information}
\usage{
get_all_spotify(input, pass, pwd = "none")
}
\arguments{
\item{input}{Data Frame containing a column \code{track.s.id} with \emph{Spotify} track ids.}

\item{pass}{Character Vector containing two entries: \emph{Client ID} and \emph{Client secret}. See \url{https://developer.spotify.com/documentation/web-api/concepts/authorization} for details.}
}
\value{
Data Frame with added information from the \emph{Spotify} Web API using the \pkg{spotilink} naming convention.
}
\description{
Retrieve a data frame containing information from the \emph{Spotify} API.
The result contains information on the tracks as well as corresponding albums and artists.
The \pkg{spotilink} naming convention is used.
}
