% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_audioanalysis_spotify.R
\name{get_audioanalysis_spotify}
\alias{get_audioanalysis_spotify}
\title{Get \emph{Spotify} API Audio Analysis for Tracks}
\usage{
get_audioanalysis_spotify(input, pass)
}
\arguments{
\item{input}{Data Frame containing a column \code{track.s.id} with \emph{Spotify} track ids.}

\item{pass}{Character Vector containing two entries: \emph{Client ID} and \emph{Client secret}. See \url{https://developer.spotify.com/documentation/web-api/concepts/authorization} for details.}
}
\value{
Data Frame with added information from the \emph{Spotify} Web API using the \pkg{spotilink} naming convention. For retrieving further metadata on tracks, please use the \code{\link{get_tracks_spotify}} function.
}
\description{
Retrieve a data frame containing information from the \emph{Spotify} API.
The result contains tracks' audio analysis using the  \pkg{spotilink} naming convention.
}
