% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_spotify_tracks_with_genreseeds.R
\name{get_available_genre_seeds_spotify}
\alias{get_available_genre_seeds_spotify}
\title{Get \emph{Spotify} genre seeds}
\usage{
get_available_genre_seeds_spotify()
}
\value{
character vector with all available Spotify genre seeds.
}
\description{
Get a list of all available Spotify genre seeds.
}
\examples{
get_available_genre_seeds_spotify()
}
