% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_linkage_quality.R
\name{get_critical_linkage_quality_observations}
\alias{get_critical_linkage_quality_observations}
\title{Get Critical Linkage Quality Observations to determine a threshold for filtering}
\usage{
get_critical_linkage_quality_observations(input)
}
\arguments{
\item{input}{}
}
\value{
List of data frames with critical linkage quality observations for all plattforms and linked entities.
}
\description{
Get a list of data frames for all platforms and entities (tracks, artists, albums) with
respective linking critical observations (quality < 1), arranged in descending quality.
With these data frames, thresholds for filtering can be determined manually by eye-balling.
}
\examples{
get_critical_linkage_quality_observations(testresults) #kann man nicht mit testresults testen, neue aus aktuellem projekt nehmen!
}
