% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_spotify_tracks_with_genreseeds.R
\name{get_multiple_recommendation_for_genre_seed_spotify}
\alias{get_multiple_recommendation_for_genre_seed_spotify}
\title{Get \emph{Spotify}-Recommendations for specific Genres}
\usage{
get_multiple_recommendation_for_genre_seed_spotify(
  seed,
  s_pass,
  n_iterations = 5
)
}
\arguments{
\item{seed}{\emph{Spotify} genre seed. Run \code{\link{get_available_genre_seeds_spotify}} to get a list of all available genre seeds.}

\item{s_pass}{Character Vector containing two entries: \emph{Client ID} and \emph{Client secret}. See \url{https://developer.spotify.com/documentation/web-api/concepts/authorization} for details.}

\item{n_iterations}{Number of recommendation iterations. Defaults to 5.}
}
\value{
List containing the genre seed, a data frame with recommendations, and information (vector and plot) on the number of news tracks added with each recommendation iteration.
}
\description{
Retrieve a frame with distinct \emph{Spotify} track information, as well for a specific genres through n iterations. Will also return information on the number of \emph{distinct} new tracks added with each recommendation iteration. Each recommendation iteration will retrieve 100 tracks tracks from the API.
}
