% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_naming_convention.R
\name{get_naming_convention}
\alias{get_naming_convention}
\title{Get Naming Convention}
\usage{
get_naming_convention()
}
\value{
Character vector with variable names.
}
\description{
Retrieve a character vector with the names of all variables that can be obtained using \pkg{spotilink}.
}
\examples{
get_naming_convention()
}
