% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tracks_lyrics_genius.R
\name{get_tracks_lyrics_genius}
\alias{get_tracks_lyrics_genius}
\title{Get \emph{Genius} song lyrics}
\usage{
get_tracks_lyrics_genius(input, g_token, threshold = 0.8)
}
\arguments{
\item{input}{Data Frame containing the following columns:
\itemize{
\item \code{track.s.id} \cr
with \emph{Spotify} track id,
\item \code{track.s.title} \cr
with \emph{Spotify} track title,
\item \code{artist.s.name} \cr
with \emph{Spotify} artist name.
}
It is advisable to first run \code{\link{get_all_spotify}} before running this command,
in order to have all the necessary information.}

\item{g_token}{Genius Authentification. Please refer to \url{https://genius.com/api-clients} for generating an access token.}

\item{threshold}{Floating point number between 0 and 1 indicating which elements to keep that were found using a fuzzy search.
The values correspond to the string similarity (1 - Jaro-Winkler distance) between the searched artist / album and the found name / title on \emph{Spotify}. \emph{spotilink} will only keep results where the artist name as well as the album title surpass the threshold.}
}
\description{
Get \emph{Genius} song lyrics
}
