% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tracks_musicbrainz.R
\name{get_tracks_musicbrainz}
\alias{get_tracks_musicbrainz}
\title{Get \emph{MusicBrainz} Track Information}
\usage{
get_tracks_musicbrainz(input, track_threshold = 0.8)
}
\arguments{
\item{input}{Data Frame containing the following columns:
\itemize{
\item\code{track.s.id} \cr
with \emph{Spotify} track ids,
\item \code{track.s.title} \cr
with \emph{Spotify} track name,
\item \code{track.s.firstartist.name} \cr
with \code{Spotify} name of first artist,
\item \code{track.s.isrc} \cr
with track ISRCs from \emph{Spotify}
}
It is advisable to first run \code{\link{get_tracks_spotify}} before running this command,
in order to have all the necessary information.}

\item{track_threshold}{Floating point number between 0 and 1 indicating which elements to keep that were found using a fuzzy search.
The values correspond to the string similarity (1 - Jaro-Winkler distance) between the track title on \emph{Spotify} and the found track title on \emph{Musicbrainz}.}
}
\value{
Data Frame with added information from the \emph{MusicBrainz} API using the  \pkg{spotilink} naming convention.
}
\description{
Retrieve a data frame containing information from the \emph{MusicBrainz} API.
The result contains information on tracks.
The \pkg{spotilink} naming convention is used.
}
\examples{
data <- data.frame(track.s.id = c('4ZXLWTmQFzM02hZwMiZfgS'),
                  track.s.title = c('Der Geist hilft unser Schwachheit auf, BWV 226'),
                  track.s.isrc = c('GBHNG1200003'))

get_tracks_musicbrainz(data)
}
