% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tracks_spotify.R
\name{get_tracks_spotify}
\alias{get_tracks_spotify}
\title{Get \emph{Spotify} API Information on Tracks}
\usage{
get_tracks_spotify(input, pass)
}
\arguments{
\item{input}{Data Frame containing a column \code{track.s.id} with \emph{Spotify} track ids.}

\item{pass}{Character Vector containing two entries: \emph{Client ID} and \emph{Client secret}. See \url{https://developer.spotify.com/documentation/web-api/concepts/authorization} for details.}
}
\value{
Data Frame with added information from the \emph{Spotify} Web API using the \pkg{spotilink} naming convention. Tracks' audio analysis is not returned. Please use the \link{get_audioanalysis_spotify} function to achieve that.
}
\description{
Retrieve a data frame containing information from the \emph{Spotify} API.
The result contains information on the tracks using the  \pkg{spotilink} naming convention.
}
\examples{
pass <- c('YOUR CLIENT ID', 'YOUR CLIENT SECRET')
data <- data.frame(track.s.id = c('4ZXLWTmQFzM02hZwMiZfgS'))
get_tracks_spotify(data, pass)
}
