% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_translations_genius.R
\name{get_translations_genius}
\alias{get_translations_genius}
\title{\emph{Genius} translation retrieval}
\usage{
get_translations_genius(input, g_token, language = "en")
}
\arguments{
\item{g_token}{Genius Authentification. Please refer to \url{https://genius.com/api-clients} for generating an access token.}

\item{language}{Genius language code for the desired translation (e.g., 'en' for English, 'de' for German, 'fr' for French).}
}
\value{
A new data frame containing the information of the input data frame, as well as new columns with information on the original lyrics language, available translations and a new column containing the version in the desired language, if the version exists.
}
\description{
\emph{Genius} translation retrieval
}
