% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_features.R
\name{select_features}
\alias{select_features}
\title{Select \emph{Spotilink}-Features from a data frame with results.}
\usage{
select_features(
  input,
  track_identifier = F,
  artist_identifier = F,
  release_identifier = F,
  label_identifier = F,
  linkage_variables = F,
  artist_descriptors = F,
  release_descriptors = F,
  platform_highlevel_audio_descriptors = F,
  ml_highlevel_audio_classificators = F,
  ml_highlevel_audio_descriptors = F,
  audio_track_metadata = F,
  aggregated_track_lowlevel_features = F,
  genre_descriptors = F,
  ml_genre_classificators = F,
  genre_tags = F,
  platform_popularity = F,
  cover_lyrics_explicitness = F
)
}
\arguments{
\item{input}{Input result data frame after}

\item{track_identifier, artist_identifier, release_identifier, label_identifier, linkage_variables, artist_descriptors, release_descriptors, spotify_highlevel_audio_descriptors, ml_highlevel_audio_classificators, ml_highlevel_audio_descriptors, audio_track_metadata, aggregated_track_lowlevel_features, genre_descriptors, ml_genre_classificators, genre_tags, platform_popularity, explicitness_and_lyrics}{True/False-flags for the inclusion of feature sets in the resulting data frame. Please refer to the feature documention sheet on GitHub for further details.Wil always include the track identifier \emph{track.s.id}.}
}
\value{
Data frame with selection of features specified through the parameter flags.
}
\description{
Select \emph{Spotilink}-Features from a data frame with results.
}
\examples{
select_features(testresults, track_identifier = T)
}
