% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_linkage_quality.R
\name{show_linkage_quality}
\alias{show_linkage_quality}
\title{Show Linkage Quality}
\usage{
show_linkage_quality(frame)
}
\arguments{
\item{frame}{Input data frame with linkage results.}
}
\value{
List with plots for each data base for that information is contained in the data frame.
}
\description{
Show linkage quality for each database and each quality measure that was calculated available in the input data frame.
}
\examples{
show_linkage_quality(testresults)
}
