% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{show_linkage_success_in_frame}
\alias{show_linkage_success_in_frame}
\title{Show Linkage Success}
\usage{
show_linkage_success_in_frame(frame)
}
\arguments{
\item{frame}{Input data frame with linkage results.}
}
\value{
Data frame with the relative frequencies of linked entities within the data for all databases that were linked to the original \emph{Spotify} data.
}
\description{
Show linkage success for each database that was linked to \emph{Spotify} with \code{spotilink}. The success calculated for distinct entities within the data is based on the entity that was used searched for on the data base (e.g., for \emph{Discogs} the success is calculated for distinct albums, while it is calculated for distinct tracks for \emph{Genius}). Also returns the relative frequency of complete linkage for tracks.
}
\examples{
show_linkage_success_in_frame(testresults)
}
