% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{show_quality_of_linkage_in_frame}
\alias{show_quality_of_linkage_in_frame}
\title{Show Linkage Quality}
\usage{
show_quality_of_linkage_in_frame(frame)
}
\arguments{
\item{frame}{Input data frame with linkage results.}
}
\value{
List with plots for each data base for that information is contained in the data frame.
}
\description{
Show linkage quality for each database and each quality measure that was calculated available in the input data frame.
}
\examples{
show_quality_of_linkage_in_frame(testresults)
}
