% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_albums_deezer.R
\name{get_albums_deezer}
\alias{get_albums_deezer}
\title{Get \emph{Deezer} Album Information}
\usage{
get_albums_deezer(input, threshold = 0.8)
}
\arguments{
\item{input}{Data Frame containing the following columns:
\itemize{
\item\code{album.s.id} \cr
with \emph{Spotify} album ids,
\item \code{album.s.title} \cr
with \emph{Spotify} album title,
\item \code{track.s.firstartist.name} \cr
with \code{Spotify} name of first artist,
}
It is advisable to first run \code{\link{get_tracks_spotify}} before running this command,
in order to have all the necessary information.}

\item{threshold}{Floating point number between 0 and 1 indicating which elements to keep that were found using a fuzzy search.
The values correspond to the string similarity (1 - Jaro-Winkler distance) between the track title on \emph{Spotify} and the found track title on \emph{Deezer}.}
}
\value{
Data Frame with added information from the \emph{Deezer} API using the  \pkg{spotilink} naming convention.
}
\description{
Retrieve a data frame containing information from the \emph{Deezer} API. This function searches for albums based on the album title and artist name from Spotify.
The result contains information on albums.
The \pkg{spotilink} naming convention is used.
}
