% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_musicbrainz.R
\name{get_all_musicbrainz}
\alias{get_all_musicbrainz}
\title{Get \emph{MusicBrainz} Information}
\usage{
get_all_musicbrainz(
  input,
  track_threshold = 0.8,
  album_threshold = 0.8,
  artist_threshold = 0.8
)
}
\arguments{
\item{input}{Data Frame containing the following columns:
\itemize{
\item\code{track.s.id} \cr
with \emph{Spotify} track ids,
\item \code{track.s.title} \cr
with \emph{Spotify} track name,
\item \code{track.s.isrc} \cr
with track ISRCs from \emph{Spotify},
\item \code{album.s.id} \cr
with \emph{Spotify} album id,
\item \code{album.s.title} \cr
with \emph{Spotify} album title,
\item \code{album.s.upc} \cr
with album UPCs from \emph{Spotify}.
\item \code{artist.s.id} \cr
with \emph{Spotify} artist id,
\item \code{artist.s.name} \cr
with \emph{Spotify} artist name.
}
It is advisable to first run \code{\link{get_all_spotify}} before running this command,
in order to have all the necessary information.}

\item{track_threshold, album_treshold, artist_threshold}{Floating point numbers between 0 and 1 indicating which elements to keep that were found using a fuzzy search.
The values correspond to the string similarity (1 - Jaro-Winkler distance) between
\enumerate{
\item the track title on \emph{Spotify} and the found track title on \emph{Musicbrainz}.
\item the album title on \emph{Spotify} and the found album title on \emph{Musicbrainz}.
\item the artist name on \emph{Spotify} and the found artist name on \emph{Musicbrainz}.
}}
}
\value{
Data Frame with added information from the \emph{MusicBrainz} API using the  \pkg{spotilink} naming convention.
}
\description{
Retrieve a data frame containing information from the \emph{MusicBrainz} API.
The result contains information on the tracks as well as corresponding albums and artists.
The \pkg{spotilink} naming convention is used.
}
\examples{
pass <- c('YOUR CLIENT ID', 'YOUR CLIENT SECRET')
data <- data.frame(track.s.id = c('4ZXLWTmQFzM02hZwMiZfgS'))
data <- get_all_spotify(data, pass)

get_all_musicbrainz(data)
}
